metrics/aesthetics.py
# metrics/aesthetics.py

from PIL import Image
import numpy as np

class AestheticScore:
    def __init__(self):
        pass

    def evaluate(self, image: Image) -> float:
        # Placeholder for aesthetic score computation
        # For demonstration, return a random score
        return np.random.uniform(0, 10)