requirements.txt
torch
diffusers
huggingface_hub
safetensors
Pillow
importlib-metadata
numpy
scipy
scikit-image
lpips
pytorch-fid
opencv-python
matplotlib